stucked with fetch api <--> db. Almost done spring security

+++ warehouse/warehouse/src/main/java/com/warehouse/warehouse/Exceptions/ClientNotFoundException.java
@@ -1,7 +1,7 @@
 package com.warehouse.warehouse.Exceptions;
 
 public class ClientNotFoundException extends RuntimeException{
-    public ClientNotFoundException(Integer id){
+    public ClientNotFoundException(Long id){
         super("Couldn't find client with id: " + id);
     }
 }
